Fix jest
