+++ packagesDev/graphql-codegen-relay-optimizer-plugin/__tests__/index.spec.ts
@@ -1,8 +1,9 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-const graphql_1 = require("graphql");
-const __1 = require("..");
-const testSchema = (0, graphql_1.buildSchema)(/* GraphQL */ `
+import '@graphql-codegen/testing'
+import { Types } from '@graphql-codegen/plugin-helpers'
+import { buildSchema, parse, print, ASTNode } from 'graphql'
+import { plugin } from '../src'
+
+const testSchema = buildSchema(/* GraphQL */ `
   type Avatar {
     id: ID!
     url: String!
@@ -18,29 +19,32 @@ const testSchema = (0, graphql_1.buildSchema)(/* GraphQL */ `
     user: User!
     users: [User!]!
   }
-`);
+`)
+
 it('can be called', async () => {
-    const testDocument = (0, graphql_1.parse)(/* GraphQL */ `
+  const testDocument = parse(/* GraphQL */ `
     query user {
       user {
         id
       }
     }
-  `);
-    await (0, __1.plugin)(testSchema, [{ document: testDocument }], {});
-});
+  `)
+  await plugin(testSchema, [{ document: testDocument }], {})
+})
+
 it('can be called with queries that include connection fragments', async () => {
-    const testDocument = (0, graphql_1.parse)(/* GraphQL */ `
+  const testDocument = parse(/* GraphQL */ `
     query user {
       users @connection(key: "foo") {
         id
       }
     }
-  `);
-    await (0, __1.plugin)(testSchema, [{ document: testDocument }], {});
-});
+  `)
+  await plugin(testSchema, [{ document: testDocument }], {})
+})
+
 it('can inline @argumentDefinitions/@arguments annotated fragments', async () => {
-    const fragmentDocument = (0, graphql_1.parse)(/* GraphQL */ `
+  const fragmentDocument = parse(/* GraphQL */ `
     fragment UserLogin on User
     @argumentDefinitions(
       height: { type: "Int", defaultValue: 10 }
@@ -53,19 +57,20 @@ it('can inline @argumentDefinitions/@arguments annotated fragments', async () =>
         url
       }
     }
-  `);
-    const queryDocument = (0, graphql_1.parse)(/* GraphQL */ `
+  `)
+  const queryDocument = parse(/* GraphQL */ `
     query user {
       users {
         ...UserLogin @arguments(height: 30, width: 30)
       }
     }
-  `);
-    const input = [{ document: fragmentDocument }, { document: queryDocument }];
-    await (0, __1.plugin)(testSchema, input, {});
-    const queryDoc = input.find((doc) => doc.document?.definitions[0].kind === 'OperationDefinition');
-    expect(queryDoc).toBeDefined();
-    expect((0, graphql_1.print)(queryDoc?.document)).toBeSimilarStringTo(/* GraphQL */ `
+  `)
+  const input: Types.DocumentFile[] = [{ document: fragmentDocument }, { document: queryDocument }]
+  await plugin(testSchema, input, {})
+  const queryDoc = input.find((doc) => doc.document?.definitions[0].kind === 'OperationDefinition')
+
+  expect(queryDoc).toBeDefined()
+  expect(print(queryDoc?.document as ASTNode)).toBeSimilarStringTo(/* GraphQL */ `
     query user {
       users {
         id
@@ -76,10 +81,11 @@ it('can inline @argumentDefinitions/@arguments annotated fragments', async () =>
         }
       }
     }
-  `);
-});
+  `)
+})
+
 it('handles unions with interfaces the correct way', async () => {
-    const schema = (0, graphql_1.buildSchema)(/* GraphQL */ `
+  const schema = buildSchema(/* GraphQL */ `
     type User {
       id: ID!
       login: String!
@@ -102,8 +108,9 @@ it('handles unions with interfaces the correct way', async () => {
     type Query {
       user: UserResult!
     }
-  `);
-    const queryDocument = (0, graphql_1.parse)(/* GraphQL */ `
+  `)
+
+  const queryDocument = parse(/* GraphQL */ `
     query user {
       user {
         ... on User {
@@ -115,12 +122,14 @@ it('handles unions with interfaces the correct way', async () => {
         }
       }
     }
-  `);
-    const input = [{ document: queryDocument }];
-    await (0, __1.plugin)(schema, input, {});
-    const queryDoc = input.find((doc) => doc.document?.definitions[0].kind === 'OperationDefinition');
-    expect(queryDoc).toBeDefined();
-    expect((0, graphql_1.print)(queryDoc?.document)).toBeSimilarStringTo(/* GraphQL */ `
+  `)
+
+  const input: Types.DocumentFile[] = [{ document: queryDocument }]
+  await plugin(schema, input, {})
+  const queryDoc = input.find((doc) => doc.document?.definitions[0].kind === 'OperationDefinition')
+
+  expect(queryDoc).toBeDefined()
+  expect(print(queryDoc?.document as ASTNode)).toBeSimilarStringTo(/* GraphQL */ `
     query user {
       user {
         ... on User {
@@ -132,5 +141,5 @@ it('handles unions with interfaces the correct way', async () => {
         }
       }
     }
-  `);
-});
+  `)
+})
