+++ packagesDev/graphql-codegen-relay-optimizer-plugin/package.json
@@ -22,6 +22,7 @@
   "devDependencies": {
     "@graphcommerce/eslint-config-pwa": "6.1.1-canary.5",
     "@graphcommerce/prettier-config-pwa": "6.1.1-canary.5",
+    "@graphql-codegen/testing": "2.0.2",
     "@types/jest": "29.4.0",
     "@types/relay-compiler": "8.0.3",
     "jest": "29.5.0",
